[tagger] Remove parent-child relationships
This was a very bug prone part of the tagger that caused several
overwritten tags and memory leaks. To prevent similar regressions in the
future (and allow us to simplify the tagger quite a bit in the future!),
we're removing this relationship, in favor of collecting collecting
container tags only when getting container events (instead of also when
getting pod and task events before), and getting k8s and ECS tags from
references present in the container itself (instead of the other way
around before, where a pod or task would add tags for each of its
containers).